made services and moved login to loginpage
