+++ .env
@@ -1,0 +1,3 @@
+REACT_APP_HOST = http://localhost
+REACT_APP_PORT = 8000
+REACT_APP_LOGIN_PATH = /api/login/
