+++ src/components/Routing.js
@@ -1,11 +1,9 @@
-import Cookies from 'js-cookie';
 import React, { Component } from 'react'
 import {
-    useNavigate,
     Route,
     Routes,
   } from 'react-router-dom';
-import LoginPage from '../pages/LoginPage';
+import {LoginWrapper} from '../pages/LoginPage';
 import {HomePageWrapper} from '../pages/HomePage';
 import ProtectedLogin from './ProtectedLogin';
 import ProtectedRoutes from './ProtectedRoutes';
@@ -13,52 +11,6 @@ import ProtectedRoutes from './ProtectedRoutes';
 
 
 class Routing extends Component {
-
-  constructor(props){
-    super(props);
-    this.state = {
-      error: ''
-    };
-  }
-
-
-  getOptions = credentials => {
-
-    const options = {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'Authorization': null,
-        'X-CSRFToken': Cookies.get('csrftoken')
-      },
-      body: JSON.stringify(credentials),
-    };
-
-    return options
-  }
-  
-
-
-  logUserIn = credentials => {
-    
-    let options = this.getOptions(credentials)
-
-
-    fetch('http://localhost:8000/api/login/', options)
-    .then(response => response.json())
-    .then(data => {
-      if(!data.detail){
-        window.localStorage.setItem('blogSiteUserLoggedIn',true);
-        this.props.navigator('/home');
-      }
-      else{
-        this.setState({
-          error: data.detail
-        })
-      }
-    });
-  }
 
   componentDidMount() {
     if (!window.localStorage.getItem('blogSiteUserLoggedIn'))
@@ -71,7 +23,7 @@ class Routing extends Component {
     return (
       <Routes>
           <Route element={<ProtectedLogin /> }>
-            <Route path='/login' element={<LoginPage  onLogin = {this.logUserIn} error = {this.state.error}/> }/>
+            <Route path='/login' element={<LoginWrapper /> }/>
           </Route>
             
           <Route element={<ProtectedRoutes /> }>
@@ -80,11 +32,6 @@ class Routing extends Component {
       </Routes>
     )
   }
-}
-
-export function RoutingWrapper(props){
-  const navigator = useNavigate();
-  return (<Routing navigator={navigator}></Routing>)
 }
 
 export default Routing;
