+++ src/pages/HomePage.js
@@ -1,4 +1,3 @@
-import Cookies from 'js-cookie'
 import React, { Component } from 'react'
 import { useNavigate } from 'react-router-dom';
